Mark messages that fall into the honeytrap as spam
We are getting a few more messages than usual that are spam, and this is
a reasonable way to at least identify whether they are spam so we can
filter them in our email _or_ eventually block them completely.

Clients will want to add a `accept_terms` value to their form field and
hide it via CSS (ideally by moving off-screen) so that spam robots will
still check the box.