🔧 Use bundle exec to ensure gems are scoped to the project
Turns out I had not used bundle exec as part of the release command, which
causes the Procfile to crash out when launched on local machines.

This is... not particularly fine; because it can add a bit of friction as people start
getting set up to work on the project.