🌱🔧  Seed a TeamMember when running db:seeds locally
We like our seed files to give us a decent starting point for local development,
as well as to be idempotent. This ensures that when we pick the project up
in the future, we have a useful set of data to work with on our local machine
without having to pull data from production.