🗃✅🌱🔧 Sprout Team Member Model (#26)
* 🗃 Sprout Team Member Model

- Small tweaks to scenario to make it terser
- Initial model, factories, and spec file
- Apply the migration and schema

* ✅ Validate presence and uniqueness of the team member attributes

We're relying on thoughtbot's shoulda matchers, because they're a well maintained
set of expectation handlers that allow us to fluently describ a bunch of
the validations that are built into Rails.

* 🌱🔧  Seed a TeamMember when running db:seeds locally

We like our seed files to give us a decent starting point for local development,
as well as to be idempotent. This ensures that when we pick the project up
in the future, we have a useful set of data to work with on our local machine
without having to pull data from production.