🔧✅ Sprout apiClient driver for feature tests
We rely on drivers to interact with the system under test because
that gives us a seam to implement API clients while we work forward.

This prevents the common anti-pattern where a projects API and other
implementation run faster than the documentation, api client, etc.

It's not a panacea; but it gives us the affordances to notice when
these test + document + implement  gaps begin to emerge.