First Steps for Team Member may Create Available Time Slot via Calendar Invite (#24)
* 🔧✅  Test creating an available time slot via calendar invite

See: https://github.com/zinc-collective/support/issues/13

Also updates the default test script so that it doesn't try and run the work that is
in progress when running feature tests.

* 🗃✅🌱🔧 Sprout Team Member Model (#26)

* 🗃 Sprout Team Member Model

- Small tweaks to scenario to make it terser
- Initial model, factories, and spec file
- Apply the migration and schema

* ✅ Validate presence and uniqueness of the team member attributes

We're relying on thoughtbot's shoulda matchers, because they're a well maintained
set of expectation handlers that allow us to fluently describ a bunch of
the validations that are built into Rails.

* 🌱🔧  Seed a TeamMember when running db:seeds locally

We like our seed files to give us a decent starting point for local development,
as well as to be idempotent. This ensures that when we pick the project up
in the future, we have a useful set of data to work with on our local machine
without having to pull data from production.

* 🔧✅ Sprout apiClient driver for feature tests

We rely on drivers to interact with the system under test because
that gives us a seam to implement API clients while we work forward.

This prevents the common anti-pattern where a projects API and other
implementation run faster than the documentation, api client, etc.

It's not a panacea; but it gives us the affordances to notice when
these test + document + implement  gaps begin to emerge.